Time: 0 ms (100.00%), Space: 6 MB (50.69%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -5,7 +5,7 @@ public:
             return true;
         int start =0;
         int end=num;
-        long long int mid = start+(end-start)/2;
+        long  int mid = start+(end-start)/2;
         while(start<=end)
         {
             if(mid*mid == num)
